+++ src/app/(routers)/favorites/_api/favoritesQueryKeys.ts
@@ -1,0 +1,4 @@
+export const favoritesQueryKeys = {
+  all: ['favorites'] as const,
+  list: () => [...favoritesQueryKeys.all, 'list'] as const,
+};
